add ability to seed data from the server
This configures the universal ng2 app to be able to seed data from
the server. That means that when the page is rendered on the server,
the data required for rendering is also planted on the page in
script tags.

When the browser loads the JS, it picks up the data in these script
tags. When the services re-run from the context of the browser, it
is aware that the data has already been seeded, and does not try
to load it again from the server.

Summary:
    IN SERVER:
        NodeJS -> Services load data ->
        Services dump data in initial render, only if in NodeJS

    IN BROWSER:
        Load dumped data from script tags into JS memory ->
        Invoke services -> Services check for seed data ->
        If seed data is found, avoid initial load

NOT DONE: Handling the passing of cookies in http Service in NodeJS.
This can probably be configured in the root component in the future.